Fix for re-applying subscription filter on reconnect (#229)
* Fix for re-applying subscription filter on reconnect

When the reconnect_stream function runs a subscriber is recreated with
the existing subscriber setup. This did not include the original
filter configuration leading to the filtering quietly disappearing on
reconnect.

Added filter_input to the `_Subscriber` object to track this configuration
and passed to the `create_subscriber` during `reconnect_stream`